deskulpt_plugin_macros::dispatch

Function proc_dispatch

pub fn proc_dispatch(_attr: TokenStream, item: TokenStream) -> TokenStream

Token stream processor for the #[dispatch] macro.

This clones the AST of the original function and performs the following modifications:

Get the type of the input parameter of the function and replace it with serde_json::Value. Not having an input parameter panics.

Get the return type of the function and replace it with anyhow::Result<serde_json::Value>. Not specifying an explicit return type panics.

Wrap the original function body in a block that deserializes the input, calls the original function, serializes the output, and returns it. Note that the original function must have a return type that the ? operator can be applied to.